BIG improvement to detect system
There's now a blueprint mesh that can be added to levels as instances, "BP_DetectZone". This has a "OnScanZone" function from an interface which returns an array of actors with the "BPI_Detectable" interface on it. Only got two functions so far, which either prints debug info or deletes them when called. You can There's a "BP_DetectableMesh" which implements the detectable interface which can also be added as an instance to levels.
Tie these together in the level blueprint however you want. Should also be very easy to add more functions to the interface/have different implementations of functions on different instances.
Still more to do for making the interfaces better but this works for now.

Also added an event dispatcher to BP_Button for ease of use and made the mesh used for detection "overlapp all" to make sure it didn't collide with things. (Might be just for the button used in my level though idk)

Removed two detect blueprints since they were just copies with different meshes of the first one, which can be done when creating an instance.

Might be good to turn this into a component that grabs a reference to the mesh in the future so you don't have to create things by adding a detectable mesh and then changing the look of it.