Discovered what a potential issue with the balls might be
It appears the cube mesh and sphere mesh that makes up the ball having different collision presets makes it act wonky. Will try to create another ball made up of a single mesh with one unified collision preset and have it be toggled of at default and toggle on when it's actually used.