TECHNICALLY FUNCTIONAL!
Need to change what the deactivation of the boss does since I changed how that works. Also some visuals to show the boss is dead. Make it flaccid or something?
Also should keep a list of all balls which has touched the boss this life and then call some "reward kill credit" on them which checks if they're still on the same iteration as when they did the damage and give kill credit if so. This might also need some sort of visual. I'm thinking some particle that travels to the balls given credit and then some golden aura around balls which has been given kill credit and are ready to exit the map.