Working version frfr no cap
Everything works as intended (I think?).
The only thing that I know isn't there yet is switching the goal point and respawn point of balls that has killed the boss. But that I can fix tomorrow. I just need to do something else for the rest of the evening now or I'll go absolutely insane.

Also it looks like ass, but that's also for the future