Finished with ramping amount of active minigames??
This should be working correctly-ish.
KNOWN BUG!! There's TWO DayTimer instances in the level. I do not know if both are needed and didn't dare to remove one of em. This means that we currently double up on minigame activation, so two activate on round start and two after one day.

If we want to change how the activation of minigames work we just need to move where "ActivateRandom" is called from in the DayTimer.